requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
googlesearch-python
janome